# pkg/dashboard/domain/specs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Graphify Dashboard",
    "description": "Administration Dashboard for Graphify models and data",
    "version": "1.0",
    "contact": {
      "name": "Graphify",
      "url": "https://github.com/amaury95/graphify",
      "email": "amauryuh@gmail.com"
    }
  },
...
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/account": {
      "get": {
        "summary": "GetAccount ...",
        "operationId": "AuthenticationService_GetAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetAccountResponse"
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          },
          "403": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "404": {
            "description": "Not found.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "500": {
            "description": "Server error.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "AuthenticationService"
        ]
      },
      "delete": {
        "summary": "DeleteAccount ...",
        "operationId": "AuthenticationService_DeleteAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          },
          "403": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "404": {
            "description": "Not found.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "500": {
            "description": "Server error.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "AuthenticationService"
        ]
      },
      "post": {
        "summary": "CreateAccount ...",
        "operationId": "AuthenticationService_CreateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1CreateAccountResponse"
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          },
          "403": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "404": {
            "description": "Not found.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "500": {
            "description": "Server error.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1CreateAccountRequest"
            }
          }
        ],
        "tags": [
          "AuthenticationService"
        ]
      },
      "put": {
        "summary": "UpdateAccount ...",
        "operationId": "AuthenticationService_UpdateAccount",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1UpdateAccountResponse"
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          },
          "403": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "404": {
            "description": "Not found.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "500": {
            "description": "Server error.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1UpdateAccountRequest"
            }
          }
        ],
        "tags": [
          "AuthenticationService"
        ]
      }
    },
    "/v1/login": {
      "post": {
        "summary": "Login ...",
        "operationId": "AuthenticationService_Login",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object",
              "properties": {}
            }
          },
          "400": {
            "description": "Bad request.",
            "schema": {
              "$ref": "#/definitions/v1ErrorResponse"
            }
          },
          "403": {
            "description": "Unauthorized.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "404": {
            "description": "Not found.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "500": {
            "description": "Server error.",
            "schema": {
              "$ref": "#/definitions/protobufEmpty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1LoginRequest"
            }
          }
        ],
        "tags": [
          "AuthenticationService"
        ],
        "security": []
      }
    },
...
  },
  "definitions": {
...
    "protobufEmpty": {
      "type": "object",
      "description": "service Foo {\n      rpc Bar(google.protobuf.Empty) returns (google.protobuf.Empty);\n    }",
      "title": "A generic empty message that you can re-use to avoid defining duplicated\nempty messages in your APIs. A typical example is to use it as the request\nor the response type of an API method. For instance:"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1Admin": {
      "type": "object",
      "properties": {
        "_key": {
          "type": "string",
          "title": "key represents the unique identifier of the admin"
        },
        "firstName": {
          "type": "string",
          "title": "first name of the admin"
        },
        "lastName": {
          "type": "string",
          "title": "last name of the admin"
        },
        "email": {
          "type": "string",
          "title": "unique email of the admin"
        },
        "passwordHash": {
          "type": "string",
          "format": "byte",
          "title": "password_hash is the hashed password of the admin"
        },
        "avatar": {
          "type": "string",
          "format": "byte",
          "title": "avatar image associated to the avatar"
        },
        "notes": {
          "type": "string",
          "title": "notes associated to the admin"
        },
        "role": {
          "$ref": "#/definitions/v1Role",
          "title": "role of the admin"
        },
        "scopes": {
          "type": "object",
          "additionalProperties": {
            "$ref": "#/definitions/v1Scope"
          },
          "title": "scopes that the admin contains for each model"
        }
      },
      "description": "Account ...\n\nAdmin data"
    },
    "v1CreateAccountRequest": {
      "type": "object",
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1Admin"
        }
      }
    },
    "v1CreateAccountResponse": {
      "type": "object",
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1Admin"
        }
      }
    },
...
    "v1ErrorResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "v1GetAccountResponse": {
      "type": "object",
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1Admin"
        }
      }
    },
...
    "v1LoginRequest": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "v1Role": {
      "type": "string",
      "enum": [
        "ROLE_UNSPECIFIED",
        "ROLE_OWNER",
        "ROLE_MANAGER"
      ],
      "default": "ROLE_UNSPECIFIED"
    },
    "v1Scope": {
      "type": "string",
      "enum": [
        "SCOPE_UNSPECIFIED",
        "SCOPE_READ",
        "SCOPE_WRITE"
      ],
      "default": "SCOPE_UNSPECIFIED"
    },
    "v1UpdateAccountRequest": {
      "type": "object",
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1Admin"
        }
      }
    },
    "v1UpdateAccountResponse": {
      "type": "object",
      "properties": {
        "admin": {
          "$ref": "#/definitions/v1Admin"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ]
}